refactor(server): remove dual-storage support for xcode graphs and command events
