+++ server/lib/tuist.ex
@@ -62,14 +62,9 @@ defmodule Tuist do
       QA.Workers.TestWorker,
       Apple,
       Xcode,
-      Xcode.Postgres,
-      Xcode.Postgres.XcodeGraph,
-      Xcode.Postgres.XcodeProject,
-      Xcode.Postgres.XcodeTarget,
-      Xcode.Clickhouse,
-      Xcode.Clickhouse.XcodeGraph,
-      Xcode.Clickhouse.XcodeProject,
-      Xcode.Clickhouse.XcodeTarget,
+      Xcode.XcodeGraph,
+      Xcode.XcodeProject,
+      Xcode.XcodeTarget,
       Loops,
       Namespace,
       Namespace.JWTToken,
