+++ server/lib/tuist/xcode/buffer/xcode_graph.ex
@@ -9,7 +9,7 @@ defmodule Tuist.Xcode.XcodeGraph.Buffer do
     insert_opts: insert_opts,
     fields: fields,
     encoding_types: encoding_types
-  } = Tuist.Xcode.Clickhouse.XcodeGraph.buffer_opts()
+  } = Tuist.Xcode.XcodeGraph.buffer_opts()
 
   def child_spec(opts) do
     opts =
