+++ server/lib/tuist/xcode/buffer/xcode_target.ex
@@ -9,7 +9,7 @@ defmodule Tuist.Xcode.XcodeTarget.Buffer do
     insert_opts: insert_opts,
     fields: fields,
     encoding_types: encoding_types
-  } = Tuist.Xcode.Clickhouse.XcodeTarget.buffer_opts()
+  } = Tuist.Xcode.XcodeTarget.buffer_opts()
 
   def child_spec(opts) do
     opts =
