+++ server/lib/tuist_web/controllers/runs_controller.ex
@@ -8,9 +8,9 @@ defmodule TuistWeb.RunsController do
   def download(conn, %{"run_id" => command_event_id}) do
     user = Authentication.current_user(conn)
 
-    with {:ok, command_event} <-
-           CommandEvents.get_command_event_by_id(command_event_id, preload: :project),
-         :ok <- Authorization.authorize(:run_read, user, command_event.project) do
+    with {:ok, command_event} <- CommandEvents.get_command_event_by_id(command_event_id),
+         {:ok, project} <- CommandEvents.get_project_for_command_event(command_event, preload: :account),
+         :ok <- Authorization.authorize(:run_read, user, project) do
       url = CommandEvents.generate_result_bundle_url(command_event)
 
       conn
