+++ server/test/support/tuist_test_support/cases/channel_case.ex
@@ -30,6 +30,11 @@ defmodule TuistWeb.ChannelCase do
 
   setup tags do
     TuistTestSupport.Cases.DataCase.setup_sandbox(tags)
+
+    on_exit(fn ->
+      TuistTestSupport.Utilities.truncate_clickhouse_tables()
+    end)
+
     :ok
   end
 end
