+++ server/test/support/tuist_test_support/cases/conn_case.ex
@@ -34,6 +34,11 @@ defmodule TuistTestSupport.Cases.ConnCase do
 
   setup tags do
     TuistTestSupport.Cases.DataCase.setup_sandbox(tags)
+
+    on_exit(fn ->
+      TuistTestSupport.Utilities.truncate_clickhouse_tables()
+    end)
+
     {:ok, conn: Phoenix.ConnTest.build_conn()}
   end
