+++ server/test/support/tuist_test_support/utilities.ex
@@ -17,4 +17,17 @@ defmodule TuistTestSupport.Utilities do
     Tuist.Xcode.XcodeTarget.Buffer.flush()
     result
   end
+
+  def truncate_clickhouse_tables do
+    commands = [
+      "TRUNCATE TABLE IF EXISTS command_events",
+      "TRUNCATE TABLE IF EXISTS xcode_graphs",
+      "TRUNCATE TABLE IF EXISTS xcode_projects",
+      "TRUNCATE TABLE IF EXISTS xcode_targets"
+    ]
+
+    for command <- commands do
+      Tuist.IngestRepo.query!(command)
+    end
+  end
 end
