+++ server/test/tuist/accounts_test.exs
@@ -28,9 +28,6 @@ defmodule Tuist.AccountsTest do
       }
     end)
 
-    stub(FunWithFlags, :enabled?, fn :clickhouse_events -> true end)
-    stub(Environment, :clickhouse_configured?, fn -> true end)
-
     :ok
   end
 
@@ -147,14 +144,12 @@ defmodule Tuist.AccountsTest do
       _user = %{account: %{id: account_id}} = AccountsFixtures.user_fixture()
       _project = %{id: project_id} = ProjectsFixtures.project_fixture(account_id: account_id)
 
-      with_flushed_ingestion_buffers(fn ->
-        CommandEventsFixtures.command_event_fixture(
-          project_id: project_id,
-          remote_test_target_hits: ["Core"],
-          remote_cache_target_hits: ["Kit"],
-          created_at: ~U[2025-05-17 15:27:00Z]
-        )
-      end)
+      CommandEventsFixtures.command_event_fixture(
+        project_id: project_id,
+        remote_test_target_hits: ["Core"],
+        remote_cache_target_hits: ["Kit"],
+        created_at: ~U[2025-05-17 15:27:00Z]
+      )
 
       # When
       got = Accounts.account_month_usage(account_id)
@@ -230,25 +225,23 @@ defmodule Tuist.AccountsTest do
       today = ~U[2025-01-02 23:00:00Z]
       stub(DateTime, :utc_now, fn -> today end)
 
-      with_flushed_ingestion_buffers(fn ->
-        CommandEventsFixtures.command_event_fixture(
-          name: "generate",
-          project_id: first_user_project.id,
-          user_id: first_user.id,
-          remote_cache_target_hits: ["Module"],
-          remote_test_target_hits: ["ModuleTeests"],
-          ran_at: ~U[2025-01-01 23:00:00Z]
-        )
+      CommandEventsFixtures.command_event_fixture(
+        name: "generate",
+        project_id: first_user_project.id,
+        user_id: first_user.id,
+        remote_cache_target_hits: ["Module"],
+        remote_test_target_hits: ["ModuleTeests"],
+        ran_at: ~U[2025-01-01 23:00:00Z]
+      )
 
-        CommandEventsFixtures.command_event_fixture(
-          name: "generate",
-          project_id: second_user_project.id,
-          user_id: second_user.id,
-          remote_cache_target_hits: ["Module"],
-          remote_test_target_hits: ["ModuleTeests"],
-          ran_at: ~U[2025-01-01 23:00:00Z]
-        )
-      end)
+      CommandEventsFixtures.command_event_fixture(
+        name: "generate",
+        project_id: second_user_project.id,
+        user_id: second_user.id,
+        remote_cache_target_hits: ["Module2"],
+        remote_test_target_hits: ["Module2Tests"],
+        ran_at: ~U[2025-01-01 23:00:00Z]
+      )
 
       # When
       assert [^first_account_customer_id] = Accounts.list_billable_customers()
@@ -1451,13 +1444,11 @@ defmodule Tuist.AccountsTest do
       Accounts.add_user_to_organization(user, organization)
 
       command_event =
-        with_flushed_ingestion_buffers(fn ->
-          CommandEventsFixtures.command_event_fixture(
-            name: "generate",
-            project_id: project.id,
-            user_id: user.id
-          )
-        end)
+        CommandEventsFixtures.command_event_fixture(
+          name: "generate",
+          project_id: project.id,
+          user_id: user.id
+        )
 
       Accounts.update_last_visited_project(user, project.id)
       code = Accounts.create_device_code("some-code")
