+++ server/test/tuist/runs/analytics_test.exs
@@ -2,10 +2,9 @@ defmodule Tuist.Runs.AnalyticsTest do
   use TuistTestSupport.Cases.DataCase
   use Mimic
 
-  alias Tuist.CommandEvents.Clickhouse.Event
   alias Tuist.IngestRepo
   alias Tuist.Runs.Analytics
-  alias Tuist.Xcode.Clickhouse.XcodeGraph
+  alias Tuist.Xcode.XcodeGraph
   alias TuistTestSupport.Fixtures.CommandEventsFixtures
   alias TuistTestSupport.Fixtures.ProjectsFixtures
   alias TuistTestSupport.Fixtures.RunsFixtures
@@ -375,15 +374,8 @@ defmodule Tuist.Runs.AnalyticsTest do
       CommandEventsFixtures.command_event_fixture(
         project_id: project.id,
         name: "generate",
-        duration: 1500,
-        created_at: ~N[2024-04-29 10:00:00]
-      )
-
-      CommandEventsFixtures.command_event_fixture(
-        project_id: project.id,
-        name: "generate",
-        duration: 2000,
-        created_at: ~N[2024-04-27 10:00:00]
+        duration: 3000,
+        created_at: ~N[2024-04-30 03:00:00]
       )
 
       # When
@@ -394,9 +386,9 @@ defmodule Tuist.Runs.AnalyticsTest do
         )
 
       # Then
-      assert got.values == [0, 1500.0, 1500.0]
-      assert got.trend == -25.0
-      assert got.total_average_duration == 1500
+      assert got.values == [0.0, 0.0, 2000.0]
+      assert got.trend == 0.0
+      assert got.total_average_duration == 2000.0
     end
 
     test "returns duration analytics for user runs only" do
@@ -423,8 +415,8 @@ defmodule Tuist.Runs.AnalyticsTest do
       CommandEventsFixtures.command_event_fixture(
         project_id: project.id,
         name: "generate",
-        duration: 1500,
-        created_at: ~N[2024-04-29 10:00:00],
+        duration: 3000,
+        created_at: ~N[2024-04-30 03:00:00],
         is_ci: false
       )
 
@@ -437,9 +429,9 @@ defmodule Tuist.Runs.AnalyticsTest do
         )
 
       # Then
-      assert got.values == [0, 1500.0, 2000.0]
+      assert got.values == [0.0, 0.0, 2500.0]
       assert got.trend == 0.0
-      assert got.total_average_duration == 1750.0
+      assert got.total_average_duration == 2500.0
     end
 
     test "returns runs analytics for the last 3 days" do
@@ -1141,24 +1133,9 @@ defmodule Tuist.Runs.AnalyticsTest do
   end
 
   describe "build_time_analytics/1" do
-    test "returns zeros when ClickHouse is not configured (PostgreSQL)" do
-      # Given
-      stub(Tuist.Environment, :clickhouse_configured?, fn -> false end)
-      project = ProjectsFixtures.project_fixture()
-
-      # When
-      got = Analytics.build_time_analytics(project_id: project.id)
-
-      # Then
-      assert got.total_time_saved == 0
-      assert got.total_build_time == 0
-      assert got.actual_build_time == 0
-    end
-
     test "returns build time analytics with real data" do
       # Given
       stub(Date, :utc_today, fn -> ~D[2024-04-30] end)
-      stub(Tuist.Environment, :clickhouse_configured?, fn -> true end)
 
       project = ProjectsFixtures.project_fixture()
 
@@ -1175,24 +1152,6 @@ defmodule Tuist.Runs.AnalyticsTest do
           created_at: ~N[2024-04-28 10:00:00],
           duration: 2000
         )
-
-      # Insert into ClickHouse Event table
-      IngestRepo.insert_all(Event, [
-        %{
-          id: command_event_1.id,
-          project_id: project.id,
-          duration: 1500,
-          created_at: NaiveDateTime.add(command_event_1.created_at, 0, :microsecond),
-          ran_at: NaiveDateTime.add(command_event_1.created_at, 0, :microsecond)
-        },
-        %{
-          id: command_event_2.id,
-          project_id: project.id,
-          duration: 2000,
-          created_at: NaiveDateTime.add(command_event_2.created_at, 0, :microsecond),
-          ran_at: NaiveDateTime.add(command_event_2.created_at, 0, :microsecond)
-        }
-      ])
 
       # Insert into ClickHouse XcodeGraph table
       IngestRepo.insert_all(XcodeGraph, [
@@ -1224,7 +1183,6 @@ defmodule Tuist.Runs.AnalyticsTest do
     test "handles empty results correctly" do
       # Given
       stub(Date, :utc_today, fn -> ~D[2024-04-30] end)
-      stub(Tuist.Environment, :clickhouse_configured?, fn -> true end)
       project = ProjectsFixtures.project_fixture()
 
       # When - no command events or xcode graphs exist
@@ -1239,7 +1197,6 @@ defmodule Tuist.Runs.AnalyticsTest do
     test "filters by project_id correctly" do
       # Given
       stub(Date, :utc_today, fn -> ~D[2024-04-30] end)
-      stub(Tuist.Environment, :clickhouse_configured?, fn -> true end)
       project1 = ProjectsFixtures.project_fixture()
       project2 = ProjectsFixtures.project_fixture()
 
@@ -1256,24 +1213,6 @@ defmodule Tuist.Runs.AnalyticsTest do
           duration: 2000,
           created_at: ~N[2024-04-28 10:00:00]
         )
-
-      # Insert into ClickHouse Event table
-      IngestRepo.insert_all(Event, [
-        %{
-          id: command_event_1.id,
-          project_id: project1.id,
-          duration: 1500,
-          created_at: NaiveDateTime.add(command_event_1.created_at, 0, :microsecond),
-          ran_at: NaiveDateTime.add(command_event_1.created_at, 0, :microsecond)
-        },
-        %{
-          id: command_event_2.id,
-          project_id: project2.id,
-          duration: 2000,
-          created_at: NaiveDateTime.add(command_event_2.created_at, 0, :microsecond),
-          ran_at: NaiveDateTime.add(command_event_2.created_at, 0, :microsecond)
-        }
-      ])
 
       # Insert into ClickHouse XcodeGraph table
       IngestRepo.insert_all(XcodeGraph, [
@@ -1305,7 +1244,6 @@ defmodule Tuist.Runs.AnalyticsTest do
     test "filters by is_ci correctly" do
       # Given
       stub(Date, :utc_today, fn -> ~D[2024-04-30] end)
-      stub(Tuist.Environment, :clickhouse_configured?, fn -> true end)
       project = ProjectsFixtures.project_fixture()
 
       command_event_ci =
@@ -1323,26 +1261,6 @@ defmodule Tuist.Runs.AnalyticsTest do
           is_ci: false,
           created_at: ~N[2024-04-28 10:00:00]
         )
-
-      # Insert into ClickHouse Event table
-      IngestRepo.insert_all(Event, [
-        %{
-          id: command_event_ci.id,
-          project_id: project.id,
-          duration: 1500,
-          is_ci: true,
-          created_at: NaiveDateTime.add(command_event_ci.created_at, 0, :microsecond),
-          ran_at: NaiveDateTime.add(command_event_ci.created_at, 0, :microsecond)
-        },
-        %{
-          id: command_event_local.id,
-          project_id: project.id,
-          duration: 2000,
-          is_ci: false,
-          created_at: NaiveDateTime.add(command_event_local.created_at, 0, :microsecond),
-          ran_at: NaiveDateTime.add(command_event_local.created_at, 0, :microsecond)
-        }
-      ])
 
       # Insert into ClickHouse XcodeGraph table
       IngestRepo.insert_all(XcodeGraph, [
@@ -1373,7 +1291,6 @@ defmodule Tuist.Runs.AnalyticsTest do
 
     test "handles custom date range" do
       # Given
-      stub(Tuist.Environment, :clickhouse_configured?, fn -> true end)
       project = ProjectsFixtures.project_fixture()
 
       command_event_in_range =
@@ -1390,25 +1307,6 @@ defmodule Tuist.Runs.AnalyticsTest do
           duration: 2000
         )
 
-      # Insert into ClickHouse Event table
-      IngestRepo.insert_all(Event, [
-        %{
-          id: command_event_in_range.id,
-          project_id: project.id,
-          duration: 1000,
-          created_at: NaiveDateTime.add(command_event_in_range.created_at, 0, :microsecond),
-          ran_at: NaiveDateTime.add(command_event_in_range.created_at, 0, :microsecond)
-        },
-        %{
-          id: command_event_out_of_range.id,
-          project_id: project.id,
-          duration: 2000,
-          created_at: NaiveDateTime.add(command_event_out_of_range.created_at, 0, :microsecond),
-          ran_at: NaiveDateTime.add(command_event_out_of_range.created_at, 0, :microsecond)
-        }
-      ])
-
-      # Insert into ClickHouse XcodeGraph table
       IngestRepo.insert_all(XcodeGraph, [
         %{
           id: UUIDv7.generate(),
@@ -1443,7 +1341,6 @@ defmodule Tuist.Runs.AnalyticsTest do
     test "handles nil duration events correctly" do
       # Given
       stub(Date, :utc_today, fn -> ~D[2024-04-30] end)
-      stub(Tuist.Environment, :clickhouse_configured?, fn -> true end)
       project = ProjectsFixtures.project_fixture()
 
       command_event =
@@ -1453,18 +1350,6 @@ defmodule Tuist.Runs.AnalyticsTest do
           created_at: ~N[2024-04-29 10:00:00]
         )
 
-      # Insert into ClickHouse Event table
-      IngestRepo.insert_all(Event, [
-        %{
-          id: command_event.id,
-          project_id: project.id,
-          duration: nil,
-          created_at: NaiveDateTime.add(command_event.created_at, 0, :microsecond),
-          ran_at: NaiveDateTime.add(command_event.created_at, 0, :microsecond)
-        }
-      ])
-
-      # Insert into ClickHouse XcodeGraph table
       IngestRepo.insert_all(XcodeGraph, [
         %{
           id: UUIDv7.generate(),
