+++ server/test/tuist_web/live/test_runs_live_test.exs
@@ -9,68 +9,9 @@ defmodule TuistWeb.TestRunsLiveTest do
   alias Tuist.Runs.Analytics
   alias TuistTestSupport.Fixtures.CommandEventsFixtures
 
-  describe "lists latest test runs - postgres" do
-    setup do
-      stub(Tuist.Environment, :clickhouse_configured?, fn -> false end)
-      stub(FunWithFlags, :enabled?, fn :clickhouse_events -> false end)
-      :ok
-    end
-
-    test "lists latest test runs", %{
-      conn: conn,
-      user: user,
-      organization: organization,
-      project: project
-    } do
-      # Given
-      _test_run_one =
-        with_flushed_ingestion_buffers(fn ->
-          CommandEventsFixtures.command_event_fixture(
-            project_id: project.id,
-            user_id: user.id,
-            name: "test",
-            command_arguments: ["test", "App"]
-          )
-
-          CommandEventsFixtures.command_event_fixture(
-            project_id: project.id,
-            user_id: user.id,
-            name: "test",
-            command_arguments: ["test", "AppTwo"]
-          )
-        end)
-
-      _test_run_one =
-        with_flushed_ingestion_buffers(fn ->
-          CommandEventsFixtures.command_event_fixture(
-            project_id: project.id,
-            user_id: user.id,
-            name: "test",
-            command_arguments: ["test", "App"]
-          )
-
-          CommandEventsFixtures.command_event_fixture(
-            project_id: project.id,
-            user_id: user.id,
-            name: "test",
-            command_arguments: ["test", "AppTwo"]
-          )
-        end)
-
-      # When
-      {:ok, lv, _html} = live(conn, ~p"/#{organization.account.name}/#{project.name}/tests/test-runs")
-
-      # Then
-      assert has_element?(lv, "span", "tuist test App")
-      assert has_element?(lv, "span", "tuist test AppTwo")
-    end
-  end
-
-  describe "lists latest test runs - clickhouse" do
+  describe "lists latest test runs" do
     setup do
       copy(Analytics)
-      stub(Tuist.Environment, :clickhouse_configured?, fn -> true end)
-      stub(FunWithFlags, :enabled?, fn :clickhouse_events -> true end)
 
       stub(Analytics, :runs_analytics, fn _, _, _ ->
         %{runs_per_period: %{}, dates: [], values: [], count: 0, trend: 0}
@@ -90,39 +31,33 @@ defmodule TuistWeb.TestRunsLiveTest do
       project: project
     } do
       # Given
-      _test_run_one =
-        with_flushed_ingestion_buffers(fn ->
-          CommandEventsFixtures.command_event_fixture(
-            project_id: project.id,
-            user_id: user.id,
-            name: "test",
-            command_arguments: ["test", "App"]
-          )
+      CommandEventsFixtures.command_event_fixture(
+        project_id: project.id,
+        user_id: user.id,
+        name: "test",
+        command_arguments: ["test", "App"]
+      )
 
-          CommandEventsFixtures.command_event_fixture(
-            project_id: project.id,
-            user_id: user.id,
-            name: "test",
-            command_arguments: ["test", "AppTwo"]
-          )
-        end)
+      CommandEventsFixtures.command_event_fixture(
+        project_id: project.id,
+        user_id: user.id,
+        name: "test",
+        command_arguments: ["test", "AppTwo"]
+      )
 
-      _test_run_one =
-        with_flushed_ingestion_buffers(fn ->
-          CommandEventsFixtures.command_event_fixture(
-            project_id: project.id,
-            user_id: user.id,
-            name: "test",
-            command_arguments: ["test", "App"]
-          )
+      CommandEventsFixtures.command_event_fixture(
+        project_id: project.id,
+        user_id: user.id,
+        name: "test",
+        command_arguments: ["test", "App"]
+      )
 
-          CommandEventsFixtures.command_event_fixture(
-            project_id: project.id,
-            user_id: user.id,
-            name: "test",
-            command_arguments: ["test", "AppTwo"]
-          )
-        end)
+      CommandEventsFixtures.command_event_fixture(
+        project_id: project.id,
+        user_id: user.id,
+        name: "test",
+        command_arguments: ["test", "AppTwo"]
+      )
 
       # When
       {:ok, lv, _html} = live(conn, ~p"/#{organization.account.name}/#{project.name}/tests/test-runs")
@@ -138,23 +73,21 @@ defmodule TuistWeb.TestRunsLiveTest do
       organization: organization,
       project: project
     } do
-      with_flushed_ingestion_buffers(fn ->
-        CommandEventsFixtures.command_event_fixture(
-          project_id: project.id,
-          user_id: user.id,
-          name: "test",
-          command_arguments: ["test", "PassingTest"],
-          status: :success
-        )
+      CommandEventsFixtures.command_event_fixture(
+        project_id: project.id,
+        user_id: user.id,
+        name: "test",
+        command_arguments: ["test", "PassingTest"],
+        status: :success
+      )
 
-        CommandEventsFixtures.command_event_fixture(
-          project_id: project.id,
-          user_id: user.id,
-          name: "test",
-          command_arguments: ["test", "FailingTest"],
-          status: :failure
-        )
-      end)
+      CommandEventsFixtures.command_event_fixture(
+        project_id: project.id,
+        user_id: user.id,
+        name: "test",
+        command_arguments: ["test", "FailingTest"],
+        status: :failure
+      )
 
       {:ok, lv, _html} = live(conn, ~p"/#{organization.account.name}/#{project.name}/tests/test-runs")
